# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: astromatch-api
  version: 0.6.0
  description: |
    Authoritative REST contract for astromatch. JSON success/error envelopes (`data`, `meta`, `error`);
    `traceId` on errors. Epic 1: identity, JWT session, recovery, minimum age, account deletion,
    consents, profile (birth, location, dynamics, bio, photos), onboarding completion, legal/privacy,
    place search (MVP stub), auth rate limits (per-IP sliding window; configurable).
    Epic 2: opaque ordered feed (`/feed/candidates`), session-style cards (no numeric score), swipe actions
    with stub daily quotas, persisted swipe events, per-user swipe burst rate limit, feed photo media for
    eligible profiles.
    Epic 3: mutual matches (`/matches`), text messages per match, optional `match` payload on swipe when
    reciprocal like creates a match, Expo push token registration (`/me/device/push-token`).
    Epic 4: mismatch feedback (`/feed/mismatch`) with focus (dynamic vs profile), persisted events, opaque
    feed reordering for targets with feedback—no exposed ranking rules.
    Epic 5: daily swipe quota (`/feed/quota`), bonus like credits from IAP, alignment boost and location pass
    timestamps on the user, stub/real billing validation (`/billing/*`), restore purchases.
    Epic 6: safety reports and blocks (`/safety/*`), operator moderation queue (`/operator/*`, `X-Operator-Key`),
...
    block enforcement in feed, matches list, and messaging; rate limits on report and block actions.
...
servers:
...
security: []
paths:
...
  /api/v1/billing/purchase/validate:
    post:
      summary: Validate a purchase receipt or transaction and grant entitlements (store-specific validation in production)
      operationId: billingPurchaseValidate
      tags:
        - Billing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseValidateRequest'
      responses:
        '200':
          description: Entitlements updated (or idempotent no-op for duplicate transaction id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessEnvelope'
        '400':
          description: Validation failed (`BILLING_VALIDATION_FAILED`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/billing/restore:
    post:
      summary: Restore prior purchases from the store and sync entitlements
      operationId: billingRestore
      tags:
        - Billing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRestoreRequest'
      responses:
        '200':
          description: Restore processed (idempotent per transaction id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessEnvelope'
        '400':
          description: Validation failed (`BILLING_VALIDATION_FAILED`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/safety/report:
    post:
      summary: Submit a user report (rate-limited per reporter per hour)
      operationId: safetyReport
      tags:
        - Safety
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SafetyReportRequest'
      responses:
        '200':
          description: Report stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSubmittedSuccessEnvelope'
        '400':
          description: Validation (`PROFILE_VALIDATION`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Report rate limit (`RATE_LIMITED`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/safety/block:
    post:
      summary: Block another user (rate-limited per blocker per day)
      operationId: safetyBlock
      tags:
        - Safety
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SafetyBlockRequest'
      responses:
        '200':
          description: Block recorded (idempotent if already blocked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessEnvelope'
        '400':
          description: Invalid target (`PROFILE_VALIDATION`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: Block rate limit (`RATE_LIMITED`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/safety/blocks:
    get:
      summary: List user IDs the current member has blocked
      operationId: safetyListBlocks
      tags:
        - Safety
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Block list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockListSuccessEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/safety/blocks/{blockedUserId}:
    delete:
      summary: Unblock a user
      operationId: safetyUnblock
      tags:
        - Safety
      security:
        - bearerAuth: []
      parameters:
        - name: blockedUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Unblock applied (no-op if not blocked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/operator/reports:
    get:
      summary: List open reports (operator API key)
      operationId: operatorListReports
      tags:
        - Operator
      parameters:
        - name: X-Operator-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Open reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorReportsSuccessEnvelope'
        '401':
          description: Invalid operator key (`OPERATOR_AUTH_INVALID`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/operator/reports/{reportId}/resolve:
    post:
      summary: Resolve a report with moderation outcome
      operationId: operatorResolveReport
      tags:
        - Operator
      parameters:
        - name: X-Operator-Key
          in: header
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorResolveRequest'
      responses:
        '200':
          description: Report resolved; user account status may be updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessEnvelope'
        '400':
          description: Validation (`PROFILE_VALIDATION`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Invalid operator key (`OPERATOR_AUTH_INVALID`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/matches:
    get:
      summary: List mutual matches for the current user
      operationId: listMatches
      tags:
        - Matches
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Match summaries (newest first)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchListSuccessEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/matches/{matchId}/messages:
    get:
      summary: List messages in a match thread
      operationId: listMatchMessages
      tags:
        - Matches
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Messages in chronological order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchMessagesSuccessEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not a participant (`MATCH_ACCESS_DENIED`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown match (`MATCH_NOT_FOUND`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      summary: Send a text message in a match thread
      operationId: sendMatchMessage
      tags:
        - Matches
      security:
        - bearerAuth: []
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMatchMessageBody'
      responses:
        '200':
          description: Message persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchMessageSuccessEnvelope'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not a participant (`MATCH_ACCESS_DENIED`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown match (`MATCH_NOT_FOUND`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SafetyReportRequest:
      type: object
      required:
        - reportedUserId
        - context
        - reasonCode
      properties:
        reportedUserId:
          type: string
          format: uuid
        context:
          type: string
          enum: [FEED, CHAT, MATCH]
        reasonCode:
          type: string
          maxLength: 64
        detail:
          type: string
          maxLength: 2000
          nullable: true
    ReportSubmitted:
      type: object
      required:
        - reportId
      properties:
        reportId:
          type: string
          format: uuid
    ReportSubmittedSuccessEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          $ref: '#/components/schemas/ReportSubmitted'
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          nullable: true
    SafetyBlockRequest:
      type: object
      required:
        - blockedUserId
      properties:
        blockedUserId:
          type: string
          format: uuid
    BlockList:
      type: object
      required:
        - blockedUserIds
      properties:
        blockedUserIds:
          type: array
          items:
            type: string
            format: uuid
    BlockListSuccessEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          $ref: '#/components/schemas/BlockList'
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          nullable: true
...
    OperatorReport:
      type: object
      required:
        - id
        - reporterId
        - reportedUserId
        - contextType
        - reasonCode
        - detail
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        reporterId:
          type: string
          format: uuid
        reportedUserId:
          type: string
          format: uuid
        contextType:
          type: string
        reasonCode:
          type: string
        detail:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    OperatorReportsSuccessEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OperatorReport'
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          nullable: true
    OperatorResolveRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [DISMISS, WARN, SUSPEND, BAN]
        note:
          type: string
          maxLength: 2000
          nullable: true
    EmptySuccessEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: object
          nullable: true
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          nullable: true
...
    PurchaseValidateRequest:
      type: object
      required:
        - platform
        - productId
      properties:
        platform:
          type: string
          enum:
            - ios
            - android
        productId:
          type: string
        receiptData:
          type: string
          nullable: true
        transactionId:
          type: string
          nullable: true
        destinationLabel:
          type: string
          nullable: true
          description: Required for location pass products when applicable.
    PurchaseRestoreItem:
      type: object
      required:
        - platform
        - productId
      properties:
        platform:
          type: string
          enum:
            - ios
            - android
        productId:
          type: string
        transactionId:
          type: string
          nullable: true
        receiptData:
          type: string
          nullable: true
        destinationLabel:
          type: string
          nullable: true
    PurchaseRestoreRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseRestoreItem'
...
    MatchSummary:
      type: object
      required:
        - matchId
        - otherUserId
        - otherEmail
      properties:
        matchId:
          type: string
          format: uuid
        otherUserId:
          type: string
          format: uuid
        otherEmail:
          type: string
          format: email
    MatchListSuccessEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MatchSummary'
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          nullable: true
    MatchMessage:
      type: object
      required:
        - id
        - senderId
        - body
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        body:
          type: string
        createdAt:
          type: string
          format: date-time
    MatchMessagesSuccessEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MatchMessage'
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          nullable: true
    MatchMessageSuccessEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          $ref: '#/components/schemas/MatchMessage'
        meta:
          type: object
          additionalProperties: true
        error:
          type: object
          nullable: true
    SendMatchMessageBody:
      type: object
      required:
        - body
      properties:
        body:
          type: string
          maxLength: 4000
...
    ApiError:
      type: object
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: string
          description: |
            Examples include `VALIDATION_ERROR`, `INVALID_CREDENTIALS`, `INVALID_REFRESH_TOKEN`,
            `INVALID_RESET_TOKEN`, `AGE_REQUIREMENT_NOT_MET`, `EMAIL_ALREADY_EXISTS`,
            `PROFILE_VALIDATION`, `RATE_LIMITED`, `QUOTA_EXCEEDED`, `SWIPE_TARGET_INVALID`,
            `MATCH_NOT_FOUND`, `MATCH_ACCESS_DENIED`, `BILLING_VALIDATION_FAILED`, `BLOCKED_USER`,
            `OPERATOR_AUTH_INVALID`, `ACCOUNT_RESTRICTED`, `ACCESS_DENIED`, `INTERNAL_ERROR`.
        message:
          type: string
        details:
          type: object
          nullable: true
          additionalProperties: true
        traceId:
          type: string
    ErrorEnvelope:
      type: object
      required:
        - data
        - meta
        - error
      properties:
        data:
          type: object
          nullable: true
        meta:
          type: object
          additionalProperties: true
        error:
          $ref: '#/components/schemas/ApiError'